move the search-ui dependencies out in `discover.search-ui`
