+++ src/views/discover/search-ui/index.js
@@ -1,0 +1,2 @@
+import './styles.scss'
+export { SearchView } from './search-ui'
