+++ src/views/discover/search-ui/styles.scss
@@ -1,0 +1,36 @@
+@import './sui-sass/base';
+@import 'styles/mixins';
+
+.search-root {
+  .top, .bottom, .results {
+    padding: 10px 20px;
+  }
+
+  .tools.overlay {
+    position: absolute;
+    z-index: 42;
+    @include bg-vibrant-tint($light-gray5, .2, 8px);
+
+    top: -55vh; // This hack makes sure we're on top of the map!
+    left: 20px; // or something. This is a [!hack].
+    padding: 10px 20px;
+    border-radius: 5px;
+  }
+
+  .tools > .unavailable {
+    display: none;
+  }
+
+  .results {
+    // make this container take at least some space.
+    min-height: 100px;
+
+    display: flex;
+    align-items: flex-start;
+    justify-content: space-evenly;
+  }
+
+  .grid.resources > .resources {
+    margin: auto
+  }
+}
